scripts/requirements.txt
paramiko
python-dotenv
pillow
